Add unix codecs for usec and nsec
