+++ internal/log_analysis/log_processor/pantherlog/tcodec/tcodec_test.go
@@ -29,14 +29,16 @@ import (
 )
 
 func TestUnixMilliseconds(t *testing.T) {
-	expect := time.Date(2020, 05, 24, 23, 50, 07, int(259*time.Millisecond.Nanoseconds()), time.UTC)
+	expect := time.Date(2020, 05, 24, 23, 50, 07, int(259*time.Millisecond.Nanoseconds()), time.UTC).Local()
 	actual := UnixMilliseconds(1590364207259)
-	require.Equal(t, expect, actual.UTC())
+	require.Equal(t, expect.Format(time.RFC3339Nano), actual.Format(time.RFC3339Nano))
+	require.Equal(t, expect, actual)
 }
 func TestUnixSeconds(t *testing.T) {
-	expect := time.Date(2020, 05, 24, 23, 50, 07, int(259*time.Millisecond.Nanoseconds()), time.UTC)
+	expect := time.Date(2020, 05, 24, 23, 50, 07, int(259*time.Millisecond.Nanoseconds()), time.UTC).Local()
 	actual := UnixSeconds(1590364207.259)
-	require.Equal(t, expect, actual.UTC())
+	require.Equal(t, expect.Format(time.RFC3339Nano), actual.Format(time.RFC3339Nano))
+	require.Equal(t, expect, actual)
 }
 
 func TestGlobalRegister(t *testing.T) {
